feat: add serializer for Flight model and use it for Trip serializer
The `flights` field of the `TripSerializer` is **read-only**. This means
when a client makes a GET request to /trips or /trips/:id then all
associated flights will be returned along with the trip fields. But it
is not possible to make a PATCH request to /trips/:id with the `flights`
field in the body. This would require additional complexity when saving
the nested/related Flight model. To avoid this, the Flight model must be
used for write purposes.